requirements.txt
Flask==2.0.1
flask-cors==3.0.10
gunicorn==20.1.0
boto3==1.26.137
torch==1.9.0
faiss-cpu==1.7.1
Pillow==8.3.2
python-dotenv==0.19.0
numpy==1.21.0
clip @ git+https://github.com/openai/CLIP.git
tenacity==8.0.1